swan OpenRTB adapter
